backend/requirements.txt
fastapi
dspy-ai[groq]
groq
uvicorn
langgraph
